% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseasusie.R
\name{coerce_scores}
\alias{coerce_scores}
\title{Coerce scores}
\usage{
coerce_scores(de, gs_background, from = "ENSEMBL", to = "ENTREZID")
}
\arguments{
\item{de}{a data.frame or similar}
}
\description{
Coerce scores
}
