% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{construct_geneset_matrix}
\alias{construct_geneset_matrix}
\title{Construct gene set matrix out of gene set table}
\usage{
construct_geneset_matrix(db)
}
\arguments{
\item{tbl}{a tibble that has at least two columns: \code{geneSet} and \code{gene}}
}
\value{
a sparse gene_set x genes membership matrix where
gene set and gene names stored in \code{colnames} and \code{rownames} respectively.
}
\description{
Construct gene set matrix out of gene set table
}
