% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{convert_alias_to_entrez}
\alias{convert_alias_to_entrez}
\title{Convert list of gene aliases to entrez ids
if 1:many mapping occurs pick the first entrez id listed in the AnnotationDbi organism database map}
\usage{
convert_alias_to_entrez(list, id_type = "alias", organism = "hsapiens")
}
\description{
Convert list of gene aliases to entrez ids
if 1:many mapping occurs pick the first entrez id listed in the AnnotationDbi organism database map
}
