% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseasusie.R
\name{convert_genelist2binaryvector}
\alias{convert_genelist2binaryvector}
\title{Gene List to Binary Vector
Convert a list of genes to a binary vector against a set of background genes
This function assumes the gene list is a strict subset of the background}
\usage{
convert_genelist2binaryvector(gene_list, background)
}
\arguments{
\item{gene_list}{a list of genes}

\item{background}{a list of background genes, length n}
}
\value{
a binary vector of length n indicating if the background gene is present in the gene list
}
\description{
Gene List to Binary Vector
Convert a list of genes to a binary vector against a set of background genes
This function assumes the gene list is a strict subset of the background
}
