% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plots.R
\name{enrichment_volcano}
\alias{enrichment_volcano}
\title{make a volcato plot
res is a tibble with columns oddsRatio, pFishersExact, in_cs, active_cs}
\usage{
enrichment_volcano(fit, ora)
}
\arguments{
\item{fit}{a fit logistic susie}

\item{ora}{output from \code{fit_ora}}
}
\description{
make a volcato plot
res is a tibble with columns oddsRatio, pFishersExact, in_cs, active_cs
}
