% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{fit_gsea_susie_webgestalt}
\alias{fit_gsea_susie_webgestalt}
\title{Fit logistic SuSiE using gene sets available from WebGestaltR}
\usage{
fit_gsea_susie_webgestalt(
  list,
  background,
  enrichDatabase = "geneontology_Biological_Process",
  organism = "hsapiens",
  ...
)
}
\arguments{
\item{list}{a list of interesting genes}

\item{background}{a list of background genes}

\item{enrichDatabase}{a database of genesets available in \code{WebGestaltR}, see \code{WebGestaltR::listGeneSet()}}
}
\value{
a list with three items \verb{$fit} \verb{$data} and \verb{$time}.
}
\description{
Fit logistic SuSiE using gene sets available from WebGestaltR
}
