% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_susie.R
\name{fit_logistic_susie}
\alias{fit_logistic_susie}
\title{Fit logistic SuSiE with variational inference}
\usage{
fit_logistic_susie(
  X,
  y,
  L = 10,
  V = 1,
  prior_weights = NULL,
  init.intercept = NULL,
  intercept = TRUE,
  Z = NULL,
  estimate_prior_variance = TRUE,
  share_prior_variance = FALSE,
  standardize = FALSE,
  center_X = TRUE,
  tol = 0.001,
  maxit = 1000,
  verbose = 0
)
}
\arguments{
\item{X}{a sparse matrix of covariates (n x p)}

\item{y}{a vector response (n)}

\item{L}{the number of non-zero effects}

\item{V}{the prior variance of effects, must either be length 1 or L}

\item{prior_weights}{the prior probability of selecting covariate (defaults to 1/p)}

\item{intercept}{boolean to include intercept in the model}

\item{Z}{matrix of covariates to always include in the model}

\item{standardize}{boolean to standardize \code{X} before fitting}

\item{center_X}{boolean to share \code{X} before fitting}

\item{tol}{tolerance to declare convergence}

\item{maxit}{maximum number of iterations}

\item{verbose}{print status/progress bar while fitting}

\item{init.initercept}{initialize the intercept parameter}

\item{estimate_prior_variance:}{boolean to estimate \code{V}}

\item{share_prior_variance:}{boolean to estimate a common \code{V} for all effects,
if False, estimate \code{V} for each effect seperately}
}
\description{
This is a driver function for logistic SuSiE (Sum of Single Effects Regression)
}
