% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_susie_veb.R
\name{fit_logistic_susie_veb_boost}
\alias{fit_logistic_susie_veb_boost}
\title{wrapper for VEB.boost to fit logistic susie
and tidy up output to be consistent with other susie-type outputs}
\usage{
fit_logistic_susie_veb_boost(X, y, L = 10, ...)
}
\description{
wrapper for VEB.boost to fit logistic susie
and tidy up output to be consistent with other susie-type outputs
}
