% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_regression.R
\name{fit_marginal_regression}
\alias{fit_marginal_regression}
\title{fit marginal regression to each geneset}
\usage{
fit_marginal_regression(X, y)
}
\arguments{
\item{X}{an n x p gene set matrix (genes x gene sets)}

\item{y}{a binary response vector (length n)}
}
\description{
fit marginal regression to each geneset
}
