% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_regression.R
\name{fit_residual_regression}
\alias{fit_residual_regression}
\title{fit marginal residual regression, using susie predictions as offset
report residual enrichments for gene sets \code{X}
after accounting for logistic SuSiE predictions}
\usage{
fit_residual_regression(X, y, fit)
}
\arguments{
\item{X}{a n x p gene set matrix, n (genes) must be same as susie fit}

\item{y}{the binary response/gene list}

\item{fit}{a logistic susie fit}
}
\description{
fit marginal residual regression, using susie predictions as offset
report residual enrichments for gene sets \code{X}
after accounting for logistic SuSiE predictions
}
