% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccm_ebnm.R
\name{fit_tccm_point_normal}
\alias{fit_tccm_point_normal}
\title{Fit Point-Normal}
\usage{
fit_tccm_point_normal(
  beta,
  se,
  X,
  logit = -1,
  update.f1 = FALSE,
  update.h = FALSE,
  maxit = 100,
  verbose = TRUE
)
}
\description{
Fit Point-Normal
}
