% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccm_ebnm.R
\name{fit_tccm_point_normal_susie}
\alias{fit_tccm_point_normal_susie}
\title{Fit Point Normal SuSiE}
\usage{
fit_tccm_point_normal_susie(
  beta,
  se,
  X,
  L = 10,
  update.f1 = TRUE,
  update.h = TRUE,
  maxit = 100,
  verbose = TRUE
)
}
\description{
Fit Point Normal SuSiE
}
