% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tccm_ebnm.R
\name{generate_f_pointmass}
\alias{generate_f_pointmass}
\title{density families
loglik is a function that takes data and a list of params
update is a function that takes data, responsibilities
and return params}
\usage{
generate_f_pointmass()
}
\description{
density families
loglik is a function that takes data and a list of params
update is a function that takes data, responsibilities
and return params
}
