% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseasusie.R
\name{generate_test_background_map}
\alias{generate_test_background_map}
\title{Covert Labels and restrict to test background}
\usage{
generate_test_background_map(
  study_background,
  background,
  study_id_type = "SYMBOL",
  background_id_type = "ENTREZID"
)
}
\arguments{
\item{study_background}{a list of genes (ID type \code{from})}

\item{background}{a list of genes (ID type \code{to})}

\item{study_id_type}{the gene ID type of \code{study_background} must be in \code{AnnotationDbi::columns(org.Hs.eg.db::org.Hs.eg.db)}}

\item{background_id_type}{the gene ID type of \code{background} must be in \code{AnnotationDbi::columns(org.Hs.eg.db::org.Hs.eg.db)}}
}
\description{
Covert Labels and restrict to test background
}
