% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_pipeline.R
\name{get_gene_set_summary}
\alias{get_gene_set_summary}
\title{map each gene set to the component with top alpha
report pip}
\usage{
get_gene_set_summary(res)
}
\description{
map each gene set to the component with top alpha
report pip
}
