% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_susie.R
\name{init_logistic_susie}
\alias{init_logistic_susie}
\title{save all the things that don't change here}
\usage{
init_logistic_susie(
  X,
  y,
  L = 10,
  V = 1,
  prior_weights = NULL,
  init.intercept = NULL,
  intercept = TRUE,
  Z = NULL,
  estimate_prior_variance = TRUE,
  share_prior_variance = FALSE,
  standardize = FALSE,
  center = TRUE
)
}
\description{
save all the things that don't change here
}
