% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_plots.R
\name{label_sig_enrichments}
\alias{label_sig_enrichments}
\title{apply FDR correction to marginal pvalues and label enrichments "depleted" or "enriched"}
\usage{
label_sig_enrichments(tbl, p, or)
}
\arguments{
\item{tbl}{a table containing columns corresponding to the other two arguments}

\item{p}{character name of column with p values}

\item{or}{character name of column with odds ratios/effect sizes}
}
\description{
apply FDR correction to marginal pvalues and label enrichments "depleted" or "enriched"
}
