% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets.R
\name{load.webGestalt.geneSet}
\alias{load.webGestalt.geneSet}
\title{wrapper for webGestatlR geneSet loading}
\usage{
load.webGestalt.geneSet(db = "geneontology_Biological_Process_noRedundant")
}
\description{
wrapper for webGestatlR geneSet loading
}
