% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets.R
\name{load_all_go}
\alias{load_all_go}
\title{convenient function to load all GO terms (across ontologies BP, MF, CC)}
\usage{
load_all_go()
}
\arguments{
\item{min.size}{minimum number of genes in term}
}
\description{
convenient function to load all GO terms (across ontologies BP, MF, CC)
}
