% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets.R
\name{load_all_msigdb}
\alias{load_all_msigdb}
\title{convenient function to load MSigDb C1-6}
\usage{
load_all_msigdb(min.size = 10)
}
\arguments{
\item{min.size}{minimum number of genes in term}
}
\description{
convenient function to load MSigDb C1-6
}
