% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets.R
\name{load_gene_sets}
\alias{load_gene_sets}
\title{load gene sets from various sources with uniform format}
\usage{
load_gene_sets(
  dbs = c("gobp", "gobp_nr", "c1", "c2", "c3", "c4", "c5", "c6", "pathways")
)
}
\arguments{
\item{dbs}{a list of gene sets to load
returns a list of gene set databases}
}
\description{
load gene sets from various sources with uniform format
}
