% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets.R
\name{load_msigdb_geneset_x}
\alias{load_msigdb_geneset_x}
\title{load MSigDB gene sets from \code{msigdbr} package}
\usage{
load_msigdb_geneset_x(db = "C2", min.size = 10)
}
\arguments{
\item{db}{is the MSigDB category to fetch (C1, C2, ... C6)}

\item{min.size}{remove gene sets smaller than this (default 10)}
}
\description{
load MSigDB gene sets from \code{msigdbr} package
}
