% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets.R
\name{load_pathways_genesets}
\alias{load_pathways_genesets}
\title{wrapper for loading geneSets from peters \code{pathways} package}
\usage{
load_pathways_genesets()
}
\description{
wrapper for loading geneSets from peters \code{pathways} package
}
