% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genesets.R
\name{load_webgestalt_geneset_x}
\alias{load_webgestalt_geneset_x}
\title{format WebGestaltR geneSets to standardized format}
\usage{
load_webgestalt_geneset_x(db, min.size = 10)
}
\description{
format WebGestaltR geneSets to standardized format
}
