% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{compute logistic regression with offset}
\usage{
logistic_regression(y, x, o)
}
\arguments{
\item{y}{a vector of binary responses}

\item{x}{a vector for a single covariate}

\item{o}{a vector of offsets (predictions from SuSiE)}
}
\description{
compute logistic regression with offset
}
