% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{make_cs_tbl}
\alias{make_cs_tbl}
\title{Make table summarizing credible sets}
\usage{
make_cs_tbl(
  fit,
  coverage = 0.9,
  min_log2bf_ser = log2(10),
  max_gs_per_component = as.integer(1e+09)
)
}
\description{
Make table summarizing credible sets
}
