% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseasusie.R
\name{prep_list_data}
\alias{prep_list_data}
\title{Prep gene list
Converts a list of genes to input for logistic SuSiE.
It is important that enrichment be assessed on an appropriate background,
so this function subsets to the intersection of genes in the study background and the union of genes in any gene set}
\usage{
prep_list_data(
  gene_list,
  study_background = NULL,
  study_id_type = "SYMBOL",
  db = "pathways"
)
}
\arguments{
\item{gene_list}{a list of genes}

\item{study_background}{the background on which \code{gene_list} was selected}

\item{study_id_type}{the gene ID type of gene_list, from \code{AnnotationDbi::columns(org.Hs.eg.db::org.Hs.eg.db)}}

\item{db}{the name of the pathway database(s) to use. Must be in \code{gseasusie::list_db}}
}
\value{
a list with elements X and y suitable for binary fit functions (fit_logistic_susie, fit_ora, fit_marginal_regression, etc)
}
\description{
Prep gene list
Converts a list of genes to input for logistic SuSiE.
It is important that enrichment be assessed on an appropriate background,
so this function subsets to the intersection of genes in the study background and the union of genes in any gene set
}
