% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseasusie.R
\name{prep_list_data2}
\alias{prep_list_data2}
\title{takes gene_list, study_backround, and geneset}
\usage{
prep_list_data2(
  gene_list,
  study_background = NULL,
  study_id_type = "SYMBOL",
  geneset
)
}
\description{
takes gene_list, study_backround, and geneset
}
