% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseasusie.R
\name{prep_score_data}
\alias{prep_score_data}
\title{Score Driver
Run GSEA on gene level statistics}
\usage{
prep_score_data(
  gene_scores,
  gene_scores_sd = NULL,
  study_id_type = "SYMBOL",
  db = "pathways"
)
}
\description{
Score Driver
Run GSEA on gene level statistics
}
