% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_susie.R
\name{rel_diff}
\alias{rel_diff}
\title{compute the relative difference (to check convergence by ELBO)}
\usage{
rel_diff(v)
}
\description{
compute the relative difference (to check convergence by ELBO)
}
