% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_pipeline.R
\name{report_susie_credible_sets}
\alias{report_susie_credible_sets}
\title{Report credible set based summary of SuSiE}
\usage{
report_susie_credible_sets(
  tbl,
  target_coverage = 0.95,
  max_coverage = 0.99,
  max_sets = 10
)
}
\arguments{
\item{tbl}{output of \alias{get_credible_set_summary} to be formatted}

\item{target_coverage}{the coverage of the credible sets to be reported}

\item{max_coverage}{report SNPs that are not in the target_coverage c.s. up this value}

\item{max_sets}{maximum number of gene sets to report for a single credible set
this is useful for large credible sets}
}
\value{
A styled kable table suitable for rendering in an HTML document
}
\description{
Report credible set based summary of SuSiE
}
