% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseasusie.R
\name{subset_gene_set_matrix}
\alias{subset_gene_set_matrix}
\title{Subset gene set matrix to genes present in test_background, and discard gene sets with too few genes}
\usage{
subset_gene_set_matrix(X, test_background, min_genes = 5)
}
\description{
Subset gene set matrix to genes present in test_background, and discard gene sets with too few genes
}
