% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_regression.R
\name{susie_coef}
\alias{susie_coef}
\title{compute expected effect for each covariate}
\usage{
susie_coef(fit)
}
\arguments{
\item{fit}{a susie-type fit}
}
\description{
compute expected effect for each covariate
}
